Refactor coarseness decorator to utils
