+++ qim3d/utils/__init__.py
@@ -1,3 +1,4 @@
+from ._decorators import coarseness
 from ._doi import *
 from ._logger import log
 from ._misc import (
